Merge
2004/09/21 17:38:18-04:00 rant-central.com!roy
Adds USB networking modules to the unslung root_fs.  Note: depmod on the slug doesn't like these modules.  It thinks they are "not for this architecture".
The ELFDATAM test is failing in depmod, so it looks like a munged #define somewhere.
The modules will work, but you have to insmod manually.

BKrev: 4152fdcbuMMgiOqsXMajFWy_o-99nQ