S45mountnfs.sh in the boot sequence (sysvinit) uses the command:
mount -a -t nfs

to mount only nfs file systems in fstab.  Busybox mount accepts
these options but ignores the -t nfs and mounts everything.  This
is bad because it remounts stuff like tmpfs partitions, obscuring
any boot time setup already done in them.  The fix is easy - just
test the file system in the -a loop if it was provided.

BKrev: 428025e4utrkgNtbkt-hh9YWbpbQfw