Merge marsco.kalmiopsis:/home/nslu2/openembedded into  marsco.kalmiopsis:/home/nslu2/ref
2005/06/05 15:28:56-07:00 kalmiopsis!jbowler
A better patch for the ARMEB ABI change in stat64, after discussion on
the uclibc mailing list.  This version doesn't damage the 2.4 functionality
(the previous patch introduced precisely the inverse error in 2.4 on
armeb!)  The change is also much more localised (probably won't be accepted
by uclibc because it changes a common file, but it's a safe change since it
is protected by __ARMEB__).

BKrev: 42a37cc4H0_djfA8an3e1gVwYtYmXQ