ipkg-utils: Make it work on fedora 9.
"control.tar.gz" got listed as "`control.tar.gz" The root cause is
somewhere else in the parser. So this is a workaround that will help
to find the right file. opkg on the device should use the proper ar
executable and that should be able to extract the right files.