Adding the recipe of mamona-input-methods.
At the moment mamona-input-methods has both ecore and gtk support,
but it's compiling just the ecore-im as we don't want it to compile
everything it needs to support gtk-im.

In the future, if we want or need, we can just modify the recipe
to compile the gtk-im, or create another recipe just for this.