pam (1.0.2 and 0.79): install headers in ${includedir}/security
While building an O.S. Systems module that uses the PAM library this
problem has been found since the code was compiling fine in a lot of
different environments and failing on OE.

Checking the source has been noticed that upstream explicitely set
includedir in configure script but autotools class overrided it by
mistake. This fixes it to the desired behaviour.