task-mamona: adding cx3110x, but just for the nokia internet tablet devices
Remember, this will probably not work if you're using a custom kernel.

At the moment, if you want to use all the hardware from the tablets, you'll
need to use some proprietary parts, like network driver (umac) and dsp.

Until we get this working with 100% free software, this will be kept inside
task-mamona.

Signed-off-by: Ricardo Salveti <ricardo.salveti@openbossa.org>
Signed-off-by: Rodrigo Vivi <rodrigo.vivi@openbossa.org>