xserver-kdrive-common: move the kdrive splash to a separate package
Move the vga ppm splash from the xserver-kdrive-common package into the new
xserver-kdrive-splash-openmoko package. this makes the x root splash cleaner
and more easily pluggable. the changes to the script check for existence of
these files and don't try and use them if they don't exist, so dependency is
soft, not hard.