Busybox 1.13.2: fix config after update form 1.11.x
The config files for default and angstrom in the Busybox 1.13.2 are outdated
(from Busybox 1.9.x and 1.11.x respectively).  These config files contain
renamed, removed, or otherwise invalid options.  This was detected when after
installing this package on angstrom it was found gzip support was missing from
tar.  Both files are updated.