classes/base.bbclass: Locate the scm base path via BBPATH.
Currently, if the first item in BBFILES isn't in the main OE repository,
the build fails entirely due to a ValueError.  This isn't optimal.  Rather
than searching through BBFILES to find it, which could be slow when
BBFILES contains thousands of files, let's just find base.bbclass and use
that repository.

Also, use better methods to find the git revision and branch, and identify
the scm based on the existance of .svn/.git/_MTN to avoid calling out
subprocesses unnecessarily.

Signed-off-by: Chris Larson <clarson@mvista.com>