image.bbclass: reorder do_rootfs
Places do_rootfs function before do_install in image creation process.

This change allows multiimage builds, when one image depends on another (before
that second image creation raced with do_rootfs for the first one) and uses
it for some purpose.

Multiimage builds are useful in situations like creating an installer image
with initramfs (built in OE) and an image of the target FS (also built in OE)
or for building virtualized system images to be placed on target FS.

Signed-off-by: Koen Kooi <koen@openembedded.org>