wireless-tools.inc: Change first use of SRC_URI from += to =
Since wireless-tools.inc is included at the top of the recipe it should clear
SRC_URI.

Without this, SRC_URI contains the recipe file, which in turn gets
installed in to the work area. When using BBPATH to set BBFILES,
this causes an error on subsequent parsing due to wireless-tools.inc
does not get copied with the recipe and is not found.

Signed-off-by: Jeremy Puhlman <jpuhlman@mvista.com>
Signed-off-by: Chris Larson <clarson@mvista.com>