angstrom, exquisite, psplash, opkg, angsrtom-images, sysvinit: abstract splash
this abstracts psplash to be generic. now as long as something provides a
splashfuncs file that sysvinit (and other scripts) can source, and all the
approproate init hooks to start the splash etc. you can use psplash,
exquisitie, usplash or anything that tickles your fancy. this moves splash
toa ${SPLASH} variable to include in your image (or override). the default in
angstrom is psplash - unless you override it. opkg also runs a configure
script that cna take forever - and so this speically sends off some splash
commands (if there) to let you know the systme is alive and working (but just
busy).

this is one commit as if this breaks things you either want to fix the minor
break or totally revert the whole patch. i hope it didn't break anything.