gnumeric: update to 1.8.4
We updated it to latest stable release and droped the old versions
since noone is preferring those. Another small change is to depends on
gdk-pixbuf-csource-native otherwise the build can fail if the host
doesn't has the development files of GTK+ installed.

The only reference to a gnumeric version has been found at
preferred-om-2008-versions.inc and it referred to a missing
version. This also makes it to point to the latest one.

Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>