coreutils: don't patch the autotools files in the native recipes
If we patch autotools files in the native recipes and build them, the make
results in the make rules trying to rerun autoconf, automake, aclocal, etc,
and those don't yet exist when coreutils-native is built.  So we split out the
common target (not native) specific bits so that native recipes aren't
including the target recipes, so we can avoid applying those patches for
native.  This also avoids some duplication of target bits between the
versions.

Signed-off-by: Chris Larson <clarson@mvista.com>