libparserutils: add recipe for version 0.0.1
LibParserUtils is a library for building efficient parsers.
Currently only Netsurf browser is using it (via hubbub), so we build just
static library.

Please note that this build doesn't use external iconv(), so only
few built-in charsets are supported (UTF-8/16, windows-125x, iso-8859-*)
We may revisit this later, if there's a need.

Signed-off-by: Dmitry Artamonow <mad_soft@inbox.ru>
Acked-by: Tom Rini <trini@embeddedalley.com>