resend xchat
as Ainulindale did not yet apply it... resending it inline which is much nicer :)

Klaus 'mrmoku' Kurzmann

--
From 38eb2e60a4088d3409b4762a86d1d1fad5effc90 Mon Sep 17 00:00:00 2001
From: Klaus Kurzmann <mok@fluxnetz.de>
Date: Tue, 26 May 2009 15:36:24 +0200
Subject: [PATCH] add a patch for xchat-2.8.4 to make it build with newer versions of gtk

This patch removes the defines of GTK_DISABLE_DEPRECATED, which is
causing xchat to not build with newer gtk. It removes all of them
to be protected against deprecation of more gtk functions in the
future.

A discussion of this topic and why having GTK_DISABLE_DEPRECATED in
release tarballs can be found here:
http://bugs.gentoo.org/show_bug.cgi?id=234458

Signed-off-by: Klaus Kurzmann <mok@fluxnetz.de>