glibc-stage.inc: don't overwrite gnu/stubs.h
"make install" installs the full version. do_stage() was overwriting
it with an empty file that tricked openssh into calling setlogin()
and logging benign errors.

Also removing a couple of lines not required anymore, I think because
"make install-headers install-lib" was switched to "make install"
in e96e82cd.

Tested with 2.6.1 on x86.

Signed-off-by: Michael Smith <msmith@cbnco.com>