patch for bluetooth audio.conf
Hello all,

to get audio routing for bluetooth headsets correct on OpenMoko phones we need
the attached patch to bluez.inc which changes audio.conf to enables the
SCORouting=PCM option.

Is this something that can be done globally? Or will it break on other
hardware? If it cannot be done globally, what is the correct way to do that?
We (SHR distro) could do that with a distro override, but if it has the
possibility to break other hardware we would not want that ;) We would rather
add a machine override then. Right now bluez4 is *not* machine dependent and
the config files are included in the base package. I see two ways to do that:

a) make bluez4 machine dependent and add overrides for om-gta01/2
b) extract the config files into some bluez4-config package and provide a
different one for om-gta01/2

Thoughts?