pciutils_3.1.2.bb: Make sure lspci reads pci.ids.gz from right location
Previously this version of 'lspci' would attempt to open pci.ids, because of
incorrect paths at configure-time.  New pciutils Makefile makes use of a
'DESTDIR' flag that we can set as the staging directory, and an 'IDSDIR' flag
which holds the location of the pci.ids{.gz} files.  Also bumped PR.

Signed-off-by: Josh Mahonin <jmahonin@cbnco.com>
Signed-off-by: Koen Kooi <koen@openembedded.org>