qt4-x11-free.inc: Do not override the var set by qt4-x11-free-gles
Another attempt to fix the dependencies of qt4-x11-free-gles. Do not
make libqt4-FOO-dev depend on qt4-x11-free-gles which will drag in
more dependencies.

A better fix would be to build the OpenVG graphicsengine unconditionally
and build the screendriver out of this tree and install it as a plugin.

Do the same change to qt4-embedded.inc.