qt4: Hack out the pg_config invocation on configure
Currently the configure script of Qt is invoking pg_config
which is a native application and not staged. Even though
we pass the directory with the header files we will fail
the postgres test as pg_config from the host will tell us
something about /usr/include.

There is no need for us to check pg_config, hack it out.