mtools: don't install-info; fixes build on non-Debian hosts
The install-info installed into staging by dpkg-native looks for
/var/backups/infodir.bak and /usr/share/base-files/info.dir on
the build system. On Debian/Ubuntu this is benign, but on other
distros these files don't exist and the install fails.

The info file is still installed, we just don't try to rebuild the
nonexistent directory.

There are cleaner ways to solve this in the long run:

http://dev.openbossa.org/trac/mamona/ticket/139

Signed-off-by: Michael Smith <msmith@cbnco.com>
Signed-off-by: Holger Freyther <zecke@selfish.org>