Add task-metadata-track.bbclass, for tracking task changes to metadata.
This class uses events to capture the state of the datastore when the task
starts, and after it completes.  It diffs those captured states, and emits
messages showing which variables changed, and what their values were changed
to.

It provides a mechanism to blacklist variables you expect to change, both
globally and on a per-task basis.

Signed-off-by: Chris Larson <clarson@kergoth.com>