gcc-4.4.1: Disable libdecnumber for uclibc target.
We do not set FENV in uclibc so as a result we can
not build libdecnumber. We disable it for gcc 4.4.1
uclibc targets.

copysignl is in libm in uclibc so gcc complains about
this function in libgcc. We see if it is uclibc powerpc
softfloat then we add -lm to linker commandline. Not best
but works.

Signed-off-by: Khem Raj <raj.khem@gmail.com>