lzma: update to 4.65 and use INC_PR
Few header files has been added to ${STAGING_INCDIR}/lzma to allow
squashfs-tools to be compiled against it.

The gcc optimization flags has been droped for now since it ought to
have been fixed since the last test with the recipe, otherwise it
should be easy to be added back.

Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>