cmake.bbclass: add toolchain file for cmake compilations
Based on work done by Matthew Dombroski.
We want to generate a toolchain file for cmake compilations. This
adds a task that is performed before the configure task.

Additional defines can be made in the toolchain file directly in
the recipe file by using a do_generate_toolchain_file_append().
This is especially interesting for recipes that need to define
values for additionnal libraries.

Signed-off-by: Valentin Longchamp <valentin.longchamp@epfl.ch>