revert sane-srcrevs: add webkit-efl|gtk:
webkit gtk srcrev was moved to recipe in b34f0987f71bbefaf914415cce6bad418a3e3d06 and bumped as well.
Since the autofoo files in OE are tied to the SRCREV care needs to be taken when bumping it, this commit actually makes it go backward.

Maybe we should ban people that have recently gained commit access from poking at conf/distro for the first few weeks to avoid fuckups like this.

http://cgit.openembedded.org/cgit.cgi/openembedded/commit/?id=b34f0987f71bbefaf914415cce6bad418a3e3d06