opkg memory problem?
01234567890123456789012345678901234567890123456789012345678901234567890123456789
When OE does 'opkg-cl install $EVERYTHING', all the memory leaks in opkg
start to accumulate. My guess is that somewhere a memory allocation failed
but was not checked for failure, then everything started to blow up.

Obviously, the memory leaks should be plugged (working on it...) and memory
allocations checked for failure (hopefully all sorted in opkg svn now). But
there's still a danger that large images will exhaust the supply of memory,
causing massive thrashing if nothing else.

The following is untested and only reduces that chances of this occuring,
but perhaps it makes sense to do something like this?