linux-{,kirkwood-}2.6.31: Add upstream squashfs-4.0 LZMA patches.
Squashfs 4.0 has official LZMA in the works.

"Re: [Squashfs-devel] squashfs 4.0 lzma support"

According to the author "No, I don't anticipate the format
changing.  As far as LZMA is concerned, the format is dictated
by the LZMA code in the kernel".

An updated squashfs-tools with LZMA support will be committed
to OpenEmbedded in a seperate commit.

Signed-off-by: Leon Woestenberg <leon@sidebranch.com>