xterm_207.bb : fix problem on "SELinux" F11 based builds xterm_251.bb : fix problem on "SELinux" F11 based builds
This was not really F11 or SELinux problem, installer was checking for
the existance of /usr/bin/X11/xterm and trying to read the permissions
for that and replicate them into the OE build. A simple --disable-setuid
prevents this script being called and the bogus check on host tools
from occuring.