gSOAP
Hello,

In the gSOAP package I noticed the following patch was applied to the
native build: use-just-built-binary.patch
This patches the makefiles of gSOAP so that the correct executable is used.
However the problem originates from the fact that the makefile
variable $(SOAP) is overridden in gsoap_${PV}.bb with the following
directive:

EXTRA_OEMAKE = "SOAP=${STAGING_BINDIR_NATIVE}/soapcpp2"

Instead of patching the makefiles of gsoap, I've come up with a
simpler solution.

In the gsoap-native_*.bb files, overwrite EXTRA_OEMAKE like so:

EXTRA_OEMAKE = ""

This removes the problem with building gsoap natively and not using
the correct executable.

Please find my patch in attachment.

Regards,

Jan

Signed-off-by: Jan Paesmans <jan.paesmans@gmail.com>
Signed-off-by: Khem Raj <raj.khem@gmail.com>