qt4: Make the gles variants for Embedded and X11 provide the normal variant
This way a distribution/user can set PREFERRED_PROVIDER_qt4-embedded = "qt4-embedded-gles"
to select the GLES/OpenVG runtime. The default is qt4-embedded as GLES require a dedicated
library most of the time provided as a binary blob by the vendor requiring the user
to manually download and store it at the right place.

For X11 we already have a PREFERRED_PROVIDER setting in place and don't need to update.

Acked-by: Koen Kooi <koen@openembedded.org>