qmake_base.bbclass: Unexport STRIP from the environment
The Makefile's generated by Qt are using this variable and the
assignment of STRIP = $(OE_QMAKE_STRIP) will be overturned by
this export. One option would be to set STRIP = $OE_QMAKE_STRIP
in the bb file but this would break striping in package.bbclass.

With the unset STRIP we are able to generate proper -dbg packages
for Qt and other qmake based utilities.