base.bbclass: in base_contains, check for var existance before using it
When using base_contains() to check for a string in a variable for a,
if the variable is not defined an exception occurs.  By checking the
existance of the variable and returning false if it isn't there,
a value can be checked for a variable regardless of whether or not it
is defined.

Signed-off-by: Brian Pomerantz <bapper@mvista.com>
Signed-off-by: Chris Larson <clarson@mvista.com>
Acked-by: Khem Raj <raj.khem@gmail.com>
Acked-by: Marcin Juszkiewicz <marcin@juszkiewicz.com.pl>
Signed-off-by: Marcin Juszkiewicz <marcin@juszkiewicz.com.pl>