From: Ulf Samuelsson <ulf.samuelsson@atmel.com>
Add support for at91bootstrap-2.13.
Fixes problem with stack, breaking all boards, compared to 2.13-rc[1-7]
Updates kconfig to 2.6.32
More options to save code on SRAM constrained SAM9260.
Builds for the following at91 chips.
* afeb9260
* at91cap9adk
* at91cap9stk
* at91sam9g10ek,
* at91sam9g20ek
* at91sam9g45ek
* at91sam9g45ekes
* at91sam9rlek
* at91sam9xeek
* at91sam9m10ek
* at91sam9260ek
* at91sam9261ek
* at91sam9263ek

So far, it has only been downloaded and tested on the sam9g45ek (dataflash)
Therefore it is only default for this chip.

Signed-off-by: Ulf Samuelsson <ulf.samuelsson@atmel.com>
---