contrib/angstrom/build-feeds.sh: added nfs-utils
The unstable feed for angstrom had 1.1.2-r4.3 on the feed (dated 18 jun 2009)
This one has a broken mountd, causing mounting an nfs filesystem to fail.
The nfs-utils in head is r7 and builds and works fien for me.  However as
nfs-utils had been removed from build-feeds.sh it was not updated; therefre added it.