socat: Add version 1.7.1.2
Second try. Added package source checksums, changed PR to r0, and removed some trailing whitespace in the patch.

Allow socat 1.7.1.2 to cross build.  In configure.in, the 4th value in AC_TRY_RUN() gives a default value for cross build environments since a test program cannot be run on the build system.  There are 2 places where AC_MSG_RESULT() is used instead of a default value which seems to result in code like "#define CRDLY_SHIFT" which breaks preprocessor checks like "#if CRDLY_SHIFT >= 0".

The patch gives default values in configure and configure.in for cross compiling and checks were added in the affected source.

Signed-off-by: John Faith <john@soundmetrics.com>