linux-ea3250: New recipe
This is a kernel recipe for the LPC32XX (LPC3250) based Embedded Artists
development board. It uses public patches provided by NXP and Embedded
Artists. The other two patches provide drivers for the PWM capabilities
and fix one NXP patch.

It can be generalized for use with the PHYTEC LPC3250 development board.

Signed-off-by: Frans Meulenbroeks <fransmeulenbroeks@gmail.com>