xinit: add dependency on util-linux-ng
if util-linux-ng is build before xinit, xinit will use mcookie (exported by util-linux-ng)
Howwever if util-linux-ng is not build before xinit it will not use mcookie.
This leads to an undesired scenario. E.g. the version of xinit on the angstrom feed needed
mcookie, but as my system did not have util-linux-ng installed, startx failed.
By making the dependency explicit at least we get consisency and apparently using mcookie
is the preferred solution.

Signed-off-by: Frans Meulenbroeks <fransmeulenbroeks@gmail.com>