u-boot git: updated calamari SRCREV, add new functionality
Ths commit is for calamari (MPC8636DS) only.
It moves to the head of the mpc85xx git
I also added several patches.
These provide additional functionality w.r.t. expression handling,
As they are not calamari specific I've put them in the u-boot-git directory.
I did not want to apply them right away as I am not sure if they apply to all the various
SRCREVs that are build with this recipe and I did want to break someone else's u-boot.
The patches are also submitted upstream

Signed-off-by: Frans Meulenbroeks <fransmeulenbroeks@gmail.com>