wireshark: Upgrade from 1.0.5 to 1.2.6
Update the ieee80215.patch. The original commit message is so weak
it is not saying why the wpan_nofcs variant is to be used. I'm
respecting the original patch author here but it would be better
to have nice commit messages. My feeling tells me that we should
drop this patch as there is no indication why this patch is a good
idea.