qemu: Perform our sanity checks based on ENABLE_BINARY_LOCALE_GENERATION
Perform qemu-related checks not based on if we're ARM but based on if
we'll be using qemu for binary locale generation.  Clarify what the
first of these sanity checks does.  Next, change the check for a
provided qemu binary to be generic enough to work on all arches (and catch
distribution or user built versions of qemu).  While we're in here, correct
the gcc version check code in base.bbclass.

Signed-off-by: Tom Rini <tom_rini@mentor.com>
Acked-by: Denys Dmytriyenko <denis@denix.org>
Acked-by: Chris Larson <chris_larson@mentor.com>