eglibc-package.bbclass: Add eglibc-pic, resync eglibc-dev with glibc, bump PR
Resync what goes into eglibc-dev based on what we have in glibc-dev.  This
includes picking up the reset of the static libraries (many are missed
otherwise and static linking in SDKs fails).  We also add in a eglibc-pic
package to catch all of the pic stuff (some of which would otherwise be caught
in -dev).

Signed-off-by: Tom Rini <tom_rini@mentor.com>
Acked-by: Khem Raj <raj.khem@gmail.com>