qemu: Move gcc version check, qemu-TARGET logic into qemu.bbclass
Move the logic to determine what qemu-TARGET to run into qemu.bbclass so
we can check for the right binary in sanity.bbclass.  This code was
duplicated by glibc-package and eglibc-package anyhow and with the new fn
we can clean up the usage in these classes a bit.  Now that we have a class
for qemu stuff, and the gcc check is just for qemu, move it there.