sanity: remove /proc/sys/vm/mmap_min_addr check
Binary locale generation fails on openSUSE 11.2 and probably would fail
on some other hosts with 2.6.31+ kernels since mmap-ing page zero may be
forbidden for security reasons even with /proc/sys/vm/mmap_min_addr set
to 0 (this also affects Ubuntu 9.10, although it's been fixed there:
https://bugs.launchpad.net/ubuntu/+source/linux/+bug/423513
).

This is not a problem for recently introduced qemu 0.12.3 since it can
relocate binaries now, which is tested on openSUSE 2.6.31 kernel with
mmap_min_addr=65536 and Debian lenny 2.6.26 with mmap_min_addr set to
0, 4096, 65536.

Given that, this check is just useless. Unfortunately, it also means
that for reliable builds we have to remove qemu 0.10 which can't work
with some kernels. Patch will follow to do that.

Signed-off-by: Roman I Khimov <khimov@altell.ru>