Do not generate NEON code for qt-4.6.2 embedded on armv6
http://bugreports.qt.nokia.com/browse/QTBUG-8749
Neon code is generated for armv6 processor, where it should be armv7 only.
This results in illegal instructions with some of the qt4 libraries.

The current solution is not optimal, I just needed a quick fix for my arch
and I don't have the needed knowledge in OE to fix it cleanly, but I wanted
to point this issue out. But at least other arch (armv5 and armv4) should have
the same problem and receive the same fix.

Signed-off-by: Koen Kooi <koen@openembedded.org>