ffmpeg: moved to native staging
Old .inc file listed every single header which was to be staged.  This
causes problems when upgrading to recent 0.5.1 version or when using
older versions because headers might not exist there resp. new ones
need to be installed.

This patch converts the .inc file to new native staging method and
which calls two new functions:

* ffmpeg_create_compat_links:
  - this creates symlinks to all installed headers into the ffmpeg/
    directory. E.g.

    adler32.h -> ../libavutil/adler32.h

  - it should be probably checked whether these links are really
    needed (they are not provided by ffmpeg and other programs might
    get confused when they see new and old headers).  For now; I kept
    them.

* ffmpeg_stage_cleanup:
  - this removes data and plugin files

Signed-off-by: Enrico Scholz <enrico.scholz@sigma-chemnitz.de>
Signed-off-by: Koen Kooi <koen@openembedded.org>