openais: add new recipe
The OpenAIS Standards Based Cluster Framework is an OSI Certified
implementation of the Service Availability Forum Application
Interface Specification (AIS). The Application Interface Specification
is a software API and policies which are used to develop applications
that maintain service during faults. Restarting and failover of
applications is also provided for those deploying applications which
may not be modified.

The OpenAIS software is built to operate as plugins to the Corosync
Cluster Engine.

Signed-off-by: Roman I Khimov <khimov@altell.ru>