dbus: use $(LIBTOOL) to install 'dbus-daemon'
Depending on autotools/libtool version and configure flags, dbus-daemon
will be built sometimes below .libs/ and sometimes in base directory.

Using 'libtool --mode=install' is a portable way to deal with this.

Changing Makefile.am to use native automake methods to install this
files would be a better but more invasive fix.

Signed-off-by: Enrico Scholz <enrico.scholz@sigma-chemnitz.de>