bitbake.conf/base.bbclass: Rename the staging directory to sysroots and the populate_staging task to populate_sysroot
This change, pulled from Poky, makes the purpose of the staging directory more
obvious and the taskname more true to what it now actually does.

The layout version number is increased due to the change in layout but code to
convert existing directories and insert a symlink for backwards compatibility
is included.

This patch also includes fixes for all recipes which reference the directory.

Signed-off-by: Joshua Lock <josh@linux.intel.com>