insane.bbclass: Switch to endswith() and fix la/pc check
This changes insane.bbclass to use s.endswith("str") rather than s[-3:] ==
"str".  The primary motivation for this is that we have tests that check for
'la' and 'pc' files when we really want '.la' and '.pc' to be what we check
for.  Also, our -dev symlink test does if test1:  if test2 and test3: action,
when it could just be if test1 and test2 and test3.

Signed-off-by: Tom Rini <tom_rini@mentor.com>
Acked-by: Khem Raj <raj.khem@gmail.com>
Acked-by: Marcin Juszkiewicz <marcin@juszkiewicz.com.pl>