relocatable.bbclass: Handle files which don't have read/write permissions
It's possible to have files in our sysroot which don't have the write (or in
some cases even the read) bit set. Test for these and if they are not set
temporarily set them so that we can chrpath the binaries.

Also remove following of symlinks because if it links to a file created by the
package install we already handle it when we process the sysroot.
Further, by ignoring symlinks here we don't cause a build to fail when the
symlink is to somewhere in the host OS.

Signed-off-by: Joshua Lock <josh@linux.intel.com>