autotools.bbclass: swapped calls of intltoolize + autoreconf
Because 'intltoolize' checks whether aclocal.m4 contains a recent copy
of 'intltool.m4' and it will fail when the found version is too old,
'aclocal' must be called before 'intltoolize'.

I built projects it with intltool-0.40.6-r3, autoconf-2.65-r9,
automake-1.11.1-r3 and gettext-0.17-r6 successfully with this change but
it will require probably tests with older versions.

This patch is a requirement for
http://patchwork.openembedded.org/patch/1858/ because 'autopoint' from
the 'aclocal' call will override 'intltool' files and build will fail
then.

Signed-off-by: Enrico Scholz <enrico.scholz@sigma-chemnitz.de>
Acked-by: Chris Larson <chris_larson@mentor.com>
Acked-by: Sebastian Spaeth <Sebastian@SSpaeth.de>
Acked-by: Roman I Khimov <khimov@altell.ru>