change INHIBIT_PACKAGE_STRIP to PACKAGE_STRIP with more options
Originally approached with "package.bbclass: introduce KEEP_GNU_DEBUGLINK":
http://patchwork.openembedded.org/patch/886/

Redone with a change from INHIBIT_PACKAGE_STRIP to PACKAGE_STRIP which
allows more options. By default it does exactly the same thing as is done
today, strips the binary, but keeps the gnu.debuglink section. When set to
"no", stripping is not done at all, when set to "full" the gnu.debuglink
section is not re-added to the binary.

Signed-off-by: Roman I Khimov <khimov@altell.ru>
Acked-by: Chris Larson <chris_larson@mentor.com>