fennec hg: update SRCREVs to 1.1b2
The 16bpp patch doesn't apply anymore and is disabled. If you're using a recent pixman and X (e.g. Angstrom, SHR) then you get *almost* enough improvements to make up for the switch to 24bpp. There is a more complex fix available, but that involves patching cairo 1.9.x as well as mozilla-central. It is better to wait for cairo 1.10 to get released and the mozilla part to go intro moz-central hg.