linux-libc-headers: Delete include/scsi/scsi.h
With 2.6.31, the kernel started exporting certain SCSI headers again and
in particular scsi/scsi.h.  After talking with Christoph Hellwig and
James Bottemley (scsi maintainer) the kernel shouldn't export
scsi/scsi.h.  Removing this from the kernel, rather than trying to clean
it up to be userland parseable is the right way to go.

We bump the libc INC_PRs to make sure the end user has a sane
<scsi/scsi.h> again.

Signed-off-by: Tom Rini <tom_rini@mentor.com>
Acked-By: Koen Kooi <k-kooi@ti.com>