base.bbclass: fixed 'apply=no' behavior
After the recent change to autodetect patchfiles in SRC_URI's, files
with 'apply=no' were not copied to workdir anymore.  This broke e.g.
dropbear which marks a patch as 'apply=no' and applies it manually.

This patch evaluates content of 'apply' parameter and copies files to
workdir when it is not a patch or the parameter is false.

Slightly altered to use oe.utils -kergoth

Signed-off-by: Enrico Scholz <enrico.scholz@sigma-chemnitz.de>
Signed-off-by: Chris Larson <chris_larson@mentor.com>