encodings/font-util: swapped inter-DEPENDS
'encodings' uses font-util's pkg-config file to determine location of
the X11 font directory.  When 'font-util' is not staged yet (as it was
the case), it will use buildhost information which point to something
like /usr/share/X11 (which is outside of the staging dir) and which are
not writable.

This results into an empty 'encodings' package on first bitbake run in a
pristine tree; 'encodings' must be rebuilt so that it can use font-util
information from previous runs.

As 'font-util' itself does not require 'encodings' during build-time,
there were swapped the DEPENDS.

Signed-off-by: Enrico Scholz <enrico.scholz@sigma-chemnitz.de>
Cc: David-John Willis <John.Willis@Distant-earth.com>
Cc: Martin Jansa <Martin.Jansa@gmail.com>