rootfs_ipk/meta-toolchain: Add locking on tasks that add package_update_index_ipk
We add a lockfile of do_populate_sdk and do_rootfs (in rootfs_ipk) that is also
the lockfile that package_update_index_ipk uses.  With enough threads it is
possible that one image (or meta-toolchain) is attempting to use the package
index while another has only reached the point of generating the index leading
to an empty index being seen later on.

Acked-by: Chris Larson <chris_larson@mentor.com>
Acked-by: Roman I Khimov <khimov@altell.ru>
Signed-off-by: Tom Rini <tom_rini@mentor.com>