packaged-staging.bbclass: Change PSTAGE_PKGARCH to something dynamic
Currently, we make all pstage packages be locked to BUILD_SYS which causes
annoying warnings in some cases and doesn't allow for sharing between say 32
and 64bit (which is valid for target bits).

We now change to ${PACKAGE_ARCH}-${HOST_OS} for everything except for cross
where we need to encode HOST_SYS into the name as well.

Acked-by: Khem Raj <raj.khem@gmail.com>
Acked-by: Chris Larson <chris_larson@mentor.com>
Signed-off-by: Tom Rini <tom_rini@mentor.com>