libsndfile1_1.0.21.bb: Partially undo 21cc37782c0624d1efc9b56e94d1012557dce871
commit 21cc37782c0624d1efc9b56e94d1012557dce871
Author: Frans Meulenbroeks <fransmeulenbroeks@gmail.com>
Date:   Thu Mar 11 09:46:25 2010 +0100

    libsndfile1: updated to 1.0.21

    also fixed download location and moved two config vars to arm-common
    (not sure about ac_cv_sizeof_off_t=8, this seems odd or recipe
specific, so left that one)

Problem is that ac_cv_sys_file_offset_bits should not be cached blindly
as the same var is used by glib-2.0 to detect large file system support
if cached then glib-2.0 configure tests are not perfored and it assumes
that large file support is there. uclibc is configurable to not have
large file support compiled in hence this breaks glib-2.0 in a obscure
way hence moving back to where they were.

Signed-off-by: Khem Raj <raj.khem@gmail.com>