stagemanager-native: Expand our gunzip/tar
As we've gone over before, gunzip/tar can have a broken pipe, and
this is allowed.  In python code, we have things setup to ignore the
problem.  stage-manager-ipkg is written in shell however and we can
hit the problem there too.  There's two ways around this, one of
which is to stop checking for problems with the extraction (and hide
stderr) or we extract the archive, gunzip and then tar.  I've done
the second here.

Acked-by: Khem Raj <raj.khem@gmail.com>
Signed-off-by: Tom Rini <tom_rini@mentor.com>